[CF-43] Move base hierarchy of exceptions to cloudferrylib.base.exception
Move TimeoutException from cloudferrylib.utils.timeout_exception to
cloudferrylib.base.exception.
Remove unused ChecksumImageInvalid from cloudferrylib.utils.utils.
