+++ cloudferrylib/utils/utils.py
@@ -417,17 +417,6 @@ class UpSshTunnelClass:
         self.remove_port(self.port)
 
 
-class ChecksumImageInvalid(Exception):
-    def __init__(self, checksum_source, checksum_dest):
-        self.checksum_source = checksum_source
-        self.checksum_dest = checksum_dest
-
-    def __str__(self):
-        return repr("Checksum of image source = %s" +
-                    "Checksum of image dest = %s" %
-                    (self.checksum_source, self.checksum_dest))
-
-
 def render_info(info_values, template_path="templates",
                 template_file="info.html"):
     info_env = Environment(loader=FileSystemLoader(template_path))
